requirements.txt
streamlit
opencv-python
joblib
scikit-learn
numpy
matplotlib
Pillow
